# docs/api.yaml
openapi: 3.0.3
info:
  title: 学习计划管理平台 API
  description: Personal Learning Plan Management Platform API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nickname:
          type: string
        role:
          type: string
          enum: [admin, user]
        timezone:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, nickname]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                nickname:
                  type: string
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      refreshToken:
                        type: string
                      user:
                        $ref: '#/components/schemas/User'
        '409':
          description: Email already registered
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user info
  /courses:
    get:
      tags: [Courses]
      summary: List courses
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived, completed]
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated course list
    post:
      tags: [Courses]
      summary: Create course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, category]
              properties:
                title:
                  type: string
                category:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Course created
  /courses/{id}:
    get:
      tags: [Courses]
      summary: Get course detail with progress
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Course detail
        '404':
          description: Not found
    put:
      tags: [Courses]
      summary: Update course
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Course updated
  /courses/{id}/archive:
    put:
      tags: [Courses]
      summary: Archive course (preserves history)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Course archived
...
  /checkins:
    post:
      tags: [Check-ins]
      summary: Create check-in
      description: Returns 409 if already checked in for this item today
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planItemId]
              properties:
                planItemId:
                  type: string
                durationMinutes:
                  type: integer
                note:
                  type: string
      responses:
        '201':
          description: Check-in created
        '409':
          description: Duplicate check-in (already checked in today)
    get:
      tags: [Check-ins]
      summary: Get check-ins
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Check-in list
  /checkins/streak:
    get:
      tags: [Check-ins]
      summary: Get streak info
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current and longest streak
...
  /stats/courses:
    get:
      tags: [Statistics]
      summary: Per-course progress
      security:
        - bearerAuth: []
      parameters:
        - name: includeArchived
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Course stats
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list
        '403':
          description: Admin required
  /admin/users/{id}/role:
    put:
      tags: [Admin]
      summary: Change user role
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Role updated
  /admin/configs:
    get:
      tags: [Admin]
      summary: Get system configs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Config list
  /admin/configs/{key}:
    put:
      tags: [Admin]
      summary: Update system config
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Config updated